move : HomePostCard -> Context로 상태관리 이동
